lesson 5 finish

+++ app/Http/Controllers/Country/CountryController.php
@@ -7,6 +7,8 @@ use Illuminate\Http\Request;
 
 use App\Models\CountryModel;
 
+use Validator;
+
 class CountryController extends Controller
 {
     public function country() {
@@ -14,20 +16,50 @@ class CountryController extends Controller
     }
 
     public function countryById($id) {
-        return response()->json(CountryModel::find($id), 200);
+        $country = CountryModel::find($id);
+        if ( is_null($country) ) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
+        return response()->json($country, 200);
     }
 
     public function countrySave(Request $req) {
+        $rules = [
+            'iso' => 'required|min:2|max:2',
+            'name' => 'required|min:3',
+            'name_en' => 'required|min:3'
+        ];
+        $validator = Validator::make($req->all(), $rules);
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
         $country = CountryModel::create($req->all());
         return response()->json($country, 201);
     }
 
-    public function countryEdit(Request $req, CountryModel $country) {
+    public function countryEdit(Request $req, $id) {
+        $rules = [
+            'iso' => 'required|min:2|max:2',
+            'name' => 'required|min:3',
+            'name_en' => 'required|min:3'
+        ];
+        $validator = Validator::make($req->all(), $rules);
+        if ($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+        $country = CountryModel::find($id);
+        if ( is_null($country) ) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $country->update($req->all());
         return response()->json($country, 200);
     }
 
-    public function countryDelete(Request $req, CountryModel $country) {
+    public function countryDelete(Request $req, $id) {
+        $country = CountryModel::find($id);
+        if ( is_null($country) ) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $country->delete();
         return response()->json('', 204);
     }
